NORM() -> added the tester into a new file the tester file and changed norm
