+++ src/garbage_collector/gc_substr.c
@@ -6,33 +6,34 @@
 /*   By: lihrig <lihrig@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/04 13:21:34 by lihrig            #+#    #+#             */
-/*   Updated: 2025/05/04 13:25:08 by lihrig           ###   ########.fr       */
+/*   Updated: 2025/05/12 16:01:46 by lihrig           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
 /**
-* @brief Erstellt einen Teilstring aus einem gegebenen String, und gibt Pointer
-*			an garbage collector
-* Diese Funktion erstellt einen neuen String, der einen Teilstring
-* des übergebenen
-* Strings 's' darstellt, beginnend am Index 'start' und maximal 'len'
-* Zeichen lang.
-* Der neue String wird dynamisch alloziert. Falls 's' NULL ist, wird
-* NULL zurückgegeben. Falls 'start' größer oder gleich der Länge von 
-* 's' ist, wird ein leerer String zurückgegeben.
-* Falls 'len' plus 'start' größer als die Länge von 's' ist,
-* wird der Teilstring bis zum Ende
-* von 's' erstellt. Bei einem Speicherfehler während der
-* Allokation wird NULL zurückgegeben.
-* Der zurückgegebene String muss vom Aufrufer mit free()
-* freigegeben werden, wenn er nicht mehr benötigt wird.
-*
-* @param s Der Quellstring, aus dem der Teilstring erstellt wird
-* @param start Der Startindex für den Teilstring im Quellstring
-* @param len Die maximale Länge des zu erstellenden Teilstrings
-* @return Ein neu allozierter Teilstring oder NULL bei Fehlern
-*/
+ * @brief Erstellt einen Teilstring aus einem gegebenen String, und gibt Pointer
+ *			an garbage collector
+ * Diese Funktion erstellt einen neuen String, der einen Teilstring
+ * des übergebenen
+ * Strings 's' darstellt, beginnend am Index 'start' und maximal 'len'
+ * Zeichen lang.
+ * Der neue String wird dynamisch alloziert. Falls 's' NULL ist, wird
+ * NULL zurückgegeben. Falls 'start' größer oder gleich der Länge von
+ * 's' ist, wird ein leerer String zurückgegeben.
+ * Falls 'len' plus 'start' größer als die Länge von 's' ist,
+ * wird der Teilstring bis zum Ende
+ * von 's' erstellt. Bei einem Speicherfehler während der
+ * Allokation wird NULL zurückgegeben.
+ * Der zurückgegebene String muss vom Aufrufer mit free()
+ * freigegeben werden, wenn er nicht mehr benötigt wird.
+ *
+ * @param s Der Quellstring, aus dem der Teilstring erstellt wird
+ * @param start Der Startindex für den Teilstring im Quellstring
+ * @param len Die maximale Länge des zu erstellenden Teilstrings
+ * @return Ein neu allozierter Teilstring oder NULL bei Fehlern
+ */
 char	*gc_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*sub_s;
